Move classes to packages
