+++ CogPar/src/main/java/co/uk/cogitolearning/cogpar/CalculateValue.java
@@ -1,5 +1,7 @@
 package co.uk.cogitolearning.cogpar;
 
+import co.uk.cogitolearning.cogpar.tree.ExpressionNode;
+
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.Stack;
