+++ CogPar/src/main/java/co/uk/cogitolearning/cogpar/ParserException.java
@@ -24,6 +24,8 @@
 
 package co.uk.cogitolearning.cogpar;
 
+import co.uk.cogitolearning.cogpar.parser.Token;
+
 /**
  * A simple subclass of RuntimeException that indicates errors when trying to
  * parse the input to Parser.
