+++ CogPar/src/main/java/co/uk/cogitolearning/cogpar/tree/ExpressionNode.java
@@ -22,7 +22,7 @@
  * THE SOFTWARE.
  */
 
-package co.uk.cogitolearning.cogpar;
+package co.uk.cogitolearning.cogpar.tree;
 
 import java.util.Iterator;
