+++ CogPar/src/test/java/co/uk/cogitolearning/cogpar/ReversePolishNotationTest.java
@@ -1,5 +1,7 @@
 package co.uk.cogitolearning.cogpar;
 
+import co.uk.cogitolearning.cogpar.parser.Parser;
+import co.uk.cogitolearning.cogpar.tree.*;
 import org.junit.Assert;
 import org.junit.Test;
